Adds the youtube iframe jsapi to the modal js.
If an iframe containing a youtube video is the content in the body of a modal the iframe takes focus when a user clicks play.
Because the iframe has focus the keyboard accessibility does not function, specifically closing modal with escape key.
This update loads the youtube iframe jsapi library and targest the active player.
When a user clicks play the onStateChange event is fired and this returns focus to the primary window.
By returning focus the keyboard commands are respected.