Added an image placeholder function (#233)
* Added an image placeholder function

After placehold.it went down I wanted to standardize how we are creating
placeholder images and stop relying on external services. This patch
will create a single method of creating basic placeholder images that
utilizes the infrastructure WebDevStudios already has in place.

* Swapping function params for array of args

Swapped out the function parameters for passing in an array of args per
code review. This was done to fall in line with the coding standards for
WebDevStudios.