Added Unsplash image placeholder capabilities  (#234)
* Added an image placeholder function

After placehold.it went down I wanted to standardize how we are creating
placeholder images and stop relying on external services. This patch
will create a single method of creating basic placeholder images that
utilizes the infrastructure WebDevStudios already has in place.

* Swapping function params for array of args

Swapped out the function parameters for passing in an array of args per
code review. This was done to fall in line with the coding standards for
WebDevStudios.

* Cleaned up code with phpcs

phpcs was disabled in my editor. Whoops! This commit has the proper
cleanup as well as localizations.

* Added Unsplash image placeholder capabilities

In addition to the basic placeholder images we now have the ability to
tap into images from Unsplash.com. Unsplash provides high quality photos
that may at times be more appropriate than a basic placeholder image.

The _s_placeholder_unsplash function supports dimensions, category, and
keyword options.

Categories can be
- buildings
- food
- nature
- people
- technology
- objects

Keywords can be any comma seperated list of words