Don't enforce 4 tab size. (#340)
After discussing internally, Aubrey is correct. Setting `indent_size = 4` simply informs the editor about the "size" of the indent. Not that we should be using "4 tab width".

This would be the equivalent of forcing everyone to use the same font. Which has zero implications in version control, and is just a stylistic thing. I'm going to merge this in and give it a try.  

**To be clear** the WDS code standard is 4 tab width.

Thanks @aubreypwd