Adds new step to find/replace setup
This adds the additional line of checking for `"_s"` and replacing it with your project string, which is an instance found in the new `phpcs.xml.dist` file. Searching through the theme, this is the only spot where this string is found with double quotes so there doesn't appear to be anything that would be replaced where it shouldn't be.